:recycle::bug: move fixtures to conftest, fix reset_globals
